fix(login): redirect on all api errors
Instead of checking for 401s exclusively
we should just handle all API errors
in a way that we redirect to /login

For instance if we get a CORS error
there's no sense to show a blank
dashboard. In the future we may have
a dedicated page for other errors though.